refactor: 기타 고객 관련 API 수정 (DURI-349)
